backend/requirements.txt
fastapi
uvicorn
boto3
pandas
python-dotenv
